+++ packages/contracts/src/ape-keys.ts
@@ -1,19 +1,17 @@
 import { initContract } from "@ts-rest/core";
 import { z } from "zod";
-
 import {
   CommonResponses,
   meta,
   MonkeyResponseSchema,
   responseWithData,
-} from "@monkeytype/schemas/api";
-
-import { IdSchema } from "@monkeytype/schemas/util";
+} from "./schemas/util";
 import {
   ApeKeySchema,
   ApeKeysSchema,
   ApeKeyUserDefinedSchema,
 } from "@monkeytype/schemas/ape-keys";
+import { IdSchema } from "@monkeytype/schemas/util";
 
 export const GetApeKeyResponseSchema = responseWithData(ApeKeysSchema);
 export type GetApeKeyResponse = z.infer<typeof GetApeKeyResponseSchema>;
@@ -39,7 +37,6 @@ export const ApeKeyParamsSchema = z.object({
 export type ApeKeyParams = z.infer<typeof ApeKeyParamsSchema>;
 
 const c = initContract();
-
 export const apeKeysContract = c.router(
   {
     get: {
