+++ packages/contracts/src/presets.ts
@@ -6,7 +6,7 @@ import {
   meta,
   MonkeyResponseSchema,
   responseWithData,
-} from "@monkeytype/schemas/api";
+} from "./schemas/util";
 import {
   EditPresetRequestSchema,
   PresetSchema,
