+++ packages/contracts/src/users.ts
@@ -7,7 +7,7 @@ import {
   MonkeyResponseSchema,
   responseWithData,
   responseWithNullableData,
-} from "@monkeytype/schemas/api";
+} from "./schemas/util";
 import {
   CountByYearAndDaySchema,
   CustomThemeNameSchema,
